miscellaneous minor fixes; stat fixed for constraints satisfied

+++ src/gov/nasa/jpl/ae/solver/ConstraintLoopSolver.java
@@ -87,7 +87,7 @@ public class ConstraintLoopSolver implements Solver {
         }
       }
       for ( Integer j : deleteList ) {
-        unsatisfiedConstraints.remove( j );
+        unsatisfiedConstraints.remove( j.intValue() );
       }
       numConstrs = unsatisfiedConstraints.size();
       boolean progress = numConstrs < lastSize;
